Remove unused "node-modules" dependency
This commit removes the "node-modules" dependency from the package.json and package-lock.json files. Its usage has been deprecated and it has not been needed in the code, so its removal shouldn't affect the application's performance or functionality. The commit also takes off the dependencies and other related information of the unused module from the package-lock.json file.